fix: change addItem for veggies and co.

+++ src/main/java/com/scanly/scanlyBackend/dtos/AddOrderItemRequest.java
@@ -1,6 +1,8 @@
 package com.scanly.scanlyBackend.dtos;
 
+import java.math.BigDecimal;
+
 public record AddOrderItemRequest (
         String code,
-        int amount
+        BigDecimal amount
 ){}
